+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/DispatcherServlet.java
@@ -1,12 +1,6 @@
-package com.techcourse;
+package com.interface21.webmvc.servlet.mvc;
 
 import com.interface21.webmvc.servlet.ModelAndView;
-import com.interface21.webmvc.servlet.mvc.AnnotationHandlerAdapter;
-import com.interface21.webmvc.servlet.mvc.AnnotationHandlerMapping;
-import com.interface21.webmvc.servlet.mvc.DefaultHandlerAdapter;
-import com.interface21.webmvc.servlet.mvc.DefaultHandlerMapping;
-import com.interface21.webmvc.servlet.mvc.HandlerAdapter;
-import com.interface21.webmvc.servlet.mvc.HandlerMapping;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
@@ -28,12 +22,10 @@ public class DispatcherServlet extends HttpServlet {
 
     public DispatcherServlet() {
         this.handlerMappings = List.of(
-            new ManualHandlerMapping(),
             new AnnotationHandlerMapping(BASE_PACKAGE),
             new DefaultHandlerMapping()
         );
         this.handlerAdapters = List.of(
-            new ManualHandlerAdapter(),
             new AnnotationHandlerAdapter(),
             new DefaultHandlerAdapter()
         );
